Always use local storage for static files
Configure Django to always use local filesystem storage for static
files. This allows us to use a standard NGINX proxy setup for the static
files, which does not have the permissions issues which have been
encountered when using AWS S3 for static file storage.

This should make the AWS S3 permissions easier to work with.